[BE] refactor #130: 패키지 형식 수정
