+++ backend/d102-api/src/main/java/com/d102/api/domain/jpa/CommonCategory.java
@@ -1,6 +1,6 @@
-package com.d102.api.domain;
+package com.d102.api.domain.jpa;
 
-import com.d102.common.domain.BaseTime;
+import com.d102.common.domain.jpa.BaseTime;
 import jakarta.persistence.*;
 import lombok.*;
