+++ backend/d102-api/src/main/java/com/d102/api/repository/jpa/CommonQuestionRepository.java
@@ -1,6 +1,6 @@
 package com.d102.api.repository.jpa;
 
-import com.d102.api.domain.CommonQuestion;
+import com.d102.api.domain.jpa.CommonQuestion;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CommonQuestionRepository  extends JpaRepository<CommonQuestion, Long> {
