+++ backend/d102-api/src/main/java/com/d102/api/repository/jpa/CommonScriptRepository.java
@@ -1,6 +1,6 @@
 package com.d102.api.repository.jpa;
 
-import com.d102.api.domain.CommonScript;
+import com.d102.api.domain.jpa.CommonScript;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
