+++ backend/d102-api/src/main/java/com/d102/api/service/impl/UserServiceImpl.java
@@ -3,7 +3,7 @@ package com.d102.api.service.impl;
 import com.d102.api.dto.UserDto;
 import com.d102.api.mapper.UserMapper;
 import com.d102.api.service.UserService;
-import com.d102.common.domain.User;
+import com.d102.common.domain.jpa.User;
 import com.d102.common.exception.ExceptionType;
 import com.d102.common.exception.custom.NotFoundException;
 import com.d102.common.exception.custom.UnAuthorizeException;
