+++ backend/d102-common/src/main/java/com/d102/common/constant/RedisConstant.java
@@ -1,0 +1,10 @@
+package com.d102.common.constant;
+
+public class RedisConstant {
+
+    public static final int REFRESH_TOKEN_LIFE = 604800;
+    public static final int QUESTION_LIST_LIFE = 604800;
+    public static final int FOLLOW_UP_QUESTION_LIFE = 604800;
+    public static final int EMAIL_LIFE = 180;
+
+}
