+++ backend/d102-common/src/main/java/com/d102/common/domain/jpa/Social.java
@@ -1,4 +1,4 @@
-package com.d102.common.domain;
+package com.d102.common.domain.jpa;
 
 import jakarta.persistence.*;
 import lombok.*;
