+++ backend/d102-common/src/main/java/com/d102/common/domain/redis/FollowUpQuestionHash.java
@@ -1,5 +1,6 @@
 package com.d102.common.domain.redis;
 
+import com.d102.common.constant.RedisConstant;
 import lombok.*;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.redis.core.RedisHash;
@@ -9,7 +10,7 @@ import org.springframework.data.redis.core.RedisHash;
 @Setter
 @AllArgsConstructor
 @NoArgsConstructor
-@RedisHash(value = "followUp", timeToLive = 604800)
+@RedisHash(value = "followUp", timeToLive = RedisConstant.FOLLOW_UP_QUESTION_LIFE)
 public class FollowUpQuestionHash {
 
     @Id
