+++ backend/d102-common/src/main/java/com/d102/common/domain/redis/QuestionListHash.java
@@ -1,5 +1,6 @@
 package com.d102.common.domain.redis;
 
+import com.d102.common.constant.RedisConstant;
 import lombok.*;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.redis.core.RedisHash;
@@ -9,7 +10,7 @@ import org.springframework.data.redis.core.RedisHash;
 @Setter
 @AllArgsConstructor
 @NoArgsConstructor
-@RedisHash(value = "questionList", timeToLive = 604800)
+@RedisHash(value = "questionList", timeToLive = RedisConstant.QUESTION_LIST_LIFE)
 public class QuestionListHash {
 
     @Id
