+++ backend/d102-common/src/main/java/com/d102/common/domain/redis/RefreshTokenHash.java
@@ -1,5 +1,6 @@
 package com.d102.common.domain.redis;
 
+import com.d102.common.constant.RedisConstant;
 import lombok.*;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.redis.core.RedisHash;
@@ -9,7 +10,7 @@ import org.springframework.data.redis.core.RedisHash;
 @Setter
 @AllArgsConstructor
 @NoArgsConstructor
-@RedisHash(value = "refreshTokenHash", timeToLive = 604800)
+@RedisHash(value = "refreshTokenHash", timeToLive = RedisConstant.REFRESH_TOKEN_LIFE)
 public class RefreshTokenHash {
 
     @Id
