+++ backend/d102-common/src/main/java/com/d102/common/repository/jpa/ResumeQuestionRepository.java
@@ -1,6 +1,6 @@
 package com.d102.common.repository.jpa;
 
-import com.d102.common.domain.ResumeQuestion;
+import com.d102.common.domain.jpa.ResumeQuestion;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ResumeQuestionRepository extends JpaRepository<ResumeQuestion, Long> {
