+++ backend/d102-common/src/main/java/com/d102/common/repository/jpa/ResumeRepository.java
@@ -1,10 +1,7 @@
 package com.d102.common.repository.jpa;
 
-import com.d102.common.domain.Resume;
-import jakarta.transaction.Transactional;
+import com.d102.common.domain.jpa.Resume;
 import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
 
 public interface ResumeRepository extends JpaRepository<Resume, Long> {
