+++ backend/d102-common/src/main/java/com/d102/common/service/AsyncService.java
@@ -1,6 +1,6 @@
 package com.d102.common.service;
 
-import com.d102.common.domain.Resume;
+import com.d102.common.domain.jpa.Resume;
 
 public interface AsyncService {
