+++ backend/d102-common/src/main/java/com/d102/common/service/impl/AsyncServiceImpl.java
@@ -1,7 +1,7 @@
 package com.d102.common.service.impl;
 
-import com.d102.common.domain.Resume;
-import com.d102.common.domain.ResumeQuestion;
+import com.d102.common.domain.jpa.Resume;
+import com.d102.common.domain.jpa.ResumeQuestion;
 import com.d102.common.domain.redis.QuestionListHash;
 import com.d102.common.exception.ExceptionType;
 import com.d102.common.exception.custom.InvalidException;
