+++ backend/d102-file/src/main/java/com/d102/file/mapper/UploadMapper.java
@@ -1,7 +1,7 @@
 package com.d102.file.mapper;
 
-import com.d102.common.domain.Resume;
-import com.d102.common.domain.User;
+import com.d102.common.domain.jpa.Resume;
+import com.d102.common.domain.jpa.User;
 import com.d102.file.dto.UploadDto;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
